pyproject.toml
[tool.poetry]
name = "mlops-v1"
version = "0.1.0"
description = ""
authors = ["Ivanxiedata <ivanxiedata@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-aiplatform = "^1.75.0"
kfp = "^2.11.0"
loguru = "^0.7.3"
mlflow = "2.19.0"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.19.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
